minor testing NaturalJoinPlan, NaturalJoinScan

+++ TestProject1/IntegrationTest/NaturalJoinPlanTests.cs
@@ -91,7 +91,7 @@ namespace TestProject1.IntegrationTest
             while (res.next())
             {
                 FuzzyProbabilisticValue<string> id = res.getFieldContent<string>("DOCTOR_ID");
-                FuzzyProbabilisticValue<int> age = res.getFieldContent<int>("D_AGE");
+                FuzzyProbabilisticValue<float> age = res.getFieldContent<float>("D_AGE");
                 FuzzyProbabilisticValue<string> name = res.getFieldContent<string>("DOCTOR_NAME");
             }
         }
